+++ index.html
@@ -231,6 +231,5 @@
   </template>
   <!-- Подключение скриптов -->
   <script src="js/main.js" type="module"></script>
-  <script src="js/utils/utils.js" type="module"></script>
 </body>
 </html>
