+++ js/create-comment.js
@@ -1,29 +1,7 @@
-// Массив имён
 
-const NAMES_ARRAY = [
-  'Антуан',
-  'Роджер',
-  'Рошан',
-  'Максимильян',
-  'Хачатурян',
-  'Вазовски',
-];
-
-// Массив сообщений
-
-const MESSAGE_ARRAY = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-// Создает массив комментарий
-
-const MIN_RANDOM_ID = 1;
-const MAX_RANDOM_ID = 6;
+import { MIN_RANDOM_ID, MAX_RANDOM_ID } from './parameters.js';
+import { MESSAGE_ARRAY, NAMES_ARRAY } from './data.js';
+import { getRandomNumber } from './get-random-number.js';
 
 const createComment = (param) => new Array(param).fill(null).map(() => ({
   id: getRandomNumber(1, 25),
@@ -32,4 +10,5 @@ const createComment = (param) => new Array(param).fill(null).map(() => ({
   name: NAMES_ARRAY[getRandomNumber(1, NAMES_ARRAY.length - 1)],
 }));
 
-export {NAMES_ARRAY, MESSAGE_ARRAY, MIN_RANDOM_ID, MAX_RANDOM_ID, createComment};
+// Экспорт функции в мэйн
+export { createComment };
