+++ js/data.js
@@ -1,0 +1,40 @@
+// Массив имён
+
+const NAMES_ARRAY = [
+  'Антуан',
+  'Роджер',
+  'Рошан',
+  'Максимильян',
+  'Хачатурян',
+  'Вазовски',
+];
+
+// Массив сообщений
+
+const MESSAGE_ARRAY = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+//   Массив Фотокарточек
+
+const DESCRIPTION_ARRAY = [
+  'Кекстаграммная котейка',
+  'Блудливый котян',
+  'Ночной пуритан',
+  'Предсказуемый жрец',
+  'Развратный жмурик',
+  'Неистовый фрик',
+  'Печальный брокер',
+  'Условный знак',
+];
+
+export {
+  NAMES_ARRAY,
+  MESSAGE_ARRAY,
+  DESCRIPTION_ARRAY
+};
