+++ js/get-random-number.js
@@ -1,11 +1,9 @@
-// Переменные для получения случайного числа
-const MAX_LIKE_COUNT = 2000;
-const MIN_LIKE = 0;
+import { getRandomPositiveInt } from './utils/utils.js';
 
-// Выставляет случайное число лайков
+// Вызывает случайное число лайков
 
 const getRandomNumber = function(min, max) {
   return getRandomPositiveInt(min, max);
-}
+};
 
-export {MAX_LIKE_COUNT, MIN_LIKE, getRandomNumber};
+export { getRandomNumber };
