+++ js/main.js
@@ -1,6 +1,6 @@
-import {getRandomPositiveInt} from './utils/utils.js';
-import {MAX_LIKE_COUNT, MIN_LIKE, getRandomNumber} from './get-random-number.js';
-import {MAX_PHOTOCARD_AMOUNT, DESCRIPTION_ARRAY, setPhotocard} from './set-photocard.js';
-import {NAMES_ARRAY, MESSAGE_ARRAY, MIN_RANDOM_ID, MAX_RANDOM_ID, createComment} from './create-comment.js';
+import { setPhotocard } from './set-photocard.js';
+import { MAX_PHOTOCARD_AMOUNT } from './parameters.js';
+import './create-comment.js';
+import './get-random-number.js';
 
 setPhotocard(MAX_PHOTOCARD_AMOUNT);
