+++ js/parameters.js
@@ -1,0 +1,21 @@
+// Параметры для функции возврата случайных чисел
+
+const MIN_RANDOM_ID = 1;
+const MAX_RANDOM_ID = 6;
+
+// Переменные для получения случайного числа
+
+const MIN_LIKE = 0;
+const MAX_LIKE_COUNT = 2000;
+
+// Параметры для функции возврата фотокарточки
+
+const MAX_PHOTOCARD_AMOUNT = 25;
+
+export {
+  MIN_RANDOM_ID,
+  MAX_RANDOM_ID,
+  MIN_LIKE,
+  MAX_LIKE_COUNT,
+  MAX_PHOTOCARD_AMOUNT
+};
